Revert "Merge branch 'master' into without-dualwelding"
This reverts commit f94e8156a53e0e7555f344c8dc7f36730baa5097, reversing
changes made to c84330ec16d5efad185e18e365b3667724298dfe.

+++ neo/framework/UsercmdGen.h
@@ -126,7 +126,7 @@ public:
 	short		angles[3];						// view angles
 	signed char	forwardmove;					// forward/backward movement
 	signed char	rightmove;						// left/right movement
-	uint16		buttons;						// buttons
+	byte		buttons;						// buttons
 	int			clientGameMilliseconds;			// time this usercmd was sent from the client
 	int			serverGameMilliseconds;			// interpolated server time this was applied on
 	uint16		fireCount;						// number of times we've fired
